Ignore variables named underscore _
What
===
Ignore variables named underscore `_`.

Why
===
The variables are not accessible to other code so they can't store
state. They're more commonly used at the global level to assert that a
type satisfies an interface and are not something that causes side
effects or something that this tool should discourage.