Ignore global error variables (implements #2) (#3)
Ignore global variables that look like errors because they start with 'err' or 'Err'.

Errors are a common thing to define at the package level, and are really used more like constants. Go doesn't allow an error to be defined as a constant though, so they are technically global variables. Since they aren't used to store state and having them defined at the package level is a common pattern, it doesn't make sense to prevent their existence.

There still is leeway for improvement such as checking if the node implements the error interface.

Fixes #2